Pass scroll speed as double.

+++ tile.h
@@ -53,7 +53,7 @@ public:
     virtual ~TileManager();
     
     virtual void act();
-    virtual void draw(int scrollX, int scrollY, const Camera &);
+    virtual void draw(double scrollX, double scrollY, const Camera &);
 protected:
     int tileX;
     int tileY;
